# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Svloon",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/salon-un-availabilies": {
            "get": {
                "tags": [
                    "SalonUnAvailabily"
                ],
                "summary": "getSalonUnAvailabilyList",
                "description": "Get all SalonUnAvailabilies",
                "operationId": "cbb6be27a68614839b2583f864cfa7b5",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SalonUnAvailabily"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "SalonUnAvailabily"
                ],
                "summary": "createSalonUnAvailabily",
                "description": "Create SalonUnAvailabily",
                "operationId": "f9f4c28aa999de2cac821a85e54ececb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SalonUnAvailabily"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SalonUnAvailabily"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/salon-un-availabilies/{id}": {
            "get": {
                "tags": [
                    "SalonUnAvailabily"
                ],
                "summary": "getSalonUnAvailabilyItem",
                "description": "Get SalonUnAvailabily",
                "operationId": "446f5e9226bef04c86bad94b9bc43f94",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of SalonUnAvailabily",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SalonUnAvailabily"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "SalonUnAvailabily"
                ],
                "summary": "updateSalonUnAvailabily",
                "description": "Update SalonUnAvailabily",
                "operationId": "ddc8beb95f818e83d2128c13211ba7e2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of SalonUnAvailabily",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SalonUnAvailabily"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SalonUnAvailabily"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "SalonUnAvailabily"
                ],
                "summary": "deleteSalonUnAvailabily",
                "description": "Delete SalonUnAvailabily",
                "operationId": "d0bb9b66dc4b88c7dc51444e1cfd86d0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of SalonUnAvailabily",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SalonUnAvailabily": {
                "required": [],
                "properties": {
                    "date": {
                        "description": "",
                        "type": "string",
                        "format": "date",
                        "readOnly": false,
                        "nullable": true
                    },
                    "raison": {
                        "description": "",
                        "type": "string",
                        "readOnly": false,
                        "nullable": true
                    },
                    "created_at": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    },
                    "updated_at": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    }
}